+++ src/tools/qsettings.cpp
@@ -378,7 +378,7 @@ void QSettingsHeading::parseLine(QTextStream &stream)
     if (line[0] == QChar('[')) {
 	QString gname = line;
 
-	gname = gname.remove(0, 1);
+	gname = gname.remove((uint)0, 1);
 	if (gname[(int)gname.length() - 1] == QChar(']'))
 	    gname = gname.remove(gname.length() - 1, 1);
